# Optimise GetCustomAttribute(s)