--- a/src/coreclr/System.Private.CoreLib/src/System/Reflection/Assembly.CoreCLR.cs
+++ b/src/coreclr/System.Private.CoreLib/src/System/Reflection/Assembly.CoreCLR.cs
@@ -98,5 +98,20 @@ namespace System.Reflection
 
         [MethodImpl(MethodImplOptions.InternalCall)]
         internal static extern uint GetAssemblyCount();
+
+        internal virtual Attribute? GetCustomAttribute(Type attributeType, bool inherit)
+        {
+            // Returns an Attribute of base class/inteface attributeType on the Assembly or null if none exists.
+            // throws an AmbiguousMatchException if there are more than one defined.
+            Attribute[] attrib = Attribute.GetCustomAttributes(this, attributeType, inherit);
+
+            if (attrib == null || attrib.Length == 0)
+                return null;
+
+            if (attrib.Length == 1)
+                return attrib[0];
+
+            throw new AmbiguousMatchException(SR.RFLCT_AmbigCust);
+        }
     }
 }