… src/coreclr/System.Private.CoreLib/src/System/Attribute.CoreCLR.cs
… src/coreclr/System.Private.CoreLib/src/System/Reflection/RuntimeConstructorInfo.cs
… src/coreclr/System.Private.CoreLib/src/System/Reflection/RuntimeEventInfo.cs
… src/coreclr/System.Private.CoreLib/src/System/Reflection/RuntimeFieldInfo.cs
… src/coreclr/System.Private.CoreLib/src/System/Reflection/RuntimeMethodInfo.cs
… src/coreclr/System.Private.CoreLib/src/System/Reflection/RuntimeModule.cs
… src/coreclr/System.Private.CoreLib/src/System/Reflection/RuntimeParameterInfo.cs
… src/coreclr/System.Private.CoreLib/src/System/Reflection/RuntimePropertyInfo.cs
… src/coreclr/System.Private.CoreLib/src/System/RuntimeType.CoreCLR.cs
… src/coreclr/System.Private.CoreLib/src/System/Type.CoreCLR.cs
… src/libraries/System.Private.CoreLib/src/System/Reflection/Module.cs
… src/libraries/System.Private.CoreLib/src/System/Reflection/ParameterInfo.cs