For `GetCustomAttribute`  brings allocations down to 25% and down to 3% for  #44713 (to just the requested attribute); with a performance improvement of 25% and 56% for #44713

For `PropertyInfo`, `Type`, `ConstructorInfo`, `MethodInfo`, `FieldInfo`, `EventInfo`


|          Method |               Toolchain | Inherit | CustomAttributeVariant |       Mean | Allocated | Ratio |
|---------------- |------------------------ |-------- |----------------------- |-----------:|----------:|------:|
|    PropertyInfo | 	\master\corerun.exe |    True |     GetCustomAttribute | 1,000.4 ns |     664 B |  1.00 |
|    PropertyInfo |         \pr\corerun.exe |    True |     GetCustomAttribute |   431.2 ns |      24 B |  0.43 |
|    PropertyInfo |                     now |    True |     GetCustomAttribute |   368.5 ns |      24 B |  0.37 |
|                 | 	                    |         |                        |            |           |       |
|            Type | 	\master\corerun.exe |    True |     GetCustomAttribute |   714.6 ns |     104 B |  1.00 |
|            Type |         \pr\corerun.exe |    True |     GetCustomAttribute |   515.3 ns |      24 B |  0.72 |
|            Type |                     now |    True |     GetCustomAttribute |   450.5 ns |      24 B |  0.63 |
|                 | 	                    |         |                        |            |           |       |
| ConstructorInfo | 	\master\corerun.exe |    True |     GetCustomAttribute |   576.3 ns |     104 B |  1.00 |
| ConstructorInfo |         \pr\corerun.exe |    True |     GetCustomAttribute |   419.7 ns |      24 B |  0.73 |
| ConstructorInfo |                     now |    True |     GetCustomAttribute |   350.0 ns |      24 B |  0.61 |
|                 | 	                    |         |                        |            |           |       |
|      MethodInfo | 	\master\corerun.exe |    True |     GetCustomAttribute |   611.6 ns |     104 B |  1.00 |
|      MethodInfo |         \pr\corerun.exe |    True |     GetCustomAttribute |   462.9 ns |      24 B |  0.76 |
|      MethodInfo |                     now |    True |     GetCustomAttribute |   392.2 ns |      24 B |  0.64 |
|                 | 	                    |         |                        |            |           |       |
|       FieldInfo | 	\master\corerun.exe |    True |     GetCustomAttribute |   618.4 ns |     104 B |  1.00 |
|       FieldInfo |         \pr\corerun.exe |    True |     GetCustomAttribute |   436.7 ns |      24 B |  0.71 |
|       FieldInfo |                     now |    True |     GetCustomAttribute |   380.7 ns |      24 B |  0.61 |
|                 | 	                    |         |                        |            |           |       |
|       EventInfo | 	\master\corerun.exe |    True |     GetCustomAttribute | 1,034.5 ns |     664 B |  1.00 |
|       EventInfo |         \pr\corerun.exe |    True |     GetCustomAttribute |   471.0 ns |      24 B |  0.46 |
|       EventInfo |                     now |    True |     GetCustomAttribute |   408.2 ns |      24 B |  0.39 |

For `GetCustomAttributes` and `GetCustomAttributes(Type)` it halves allocations to just array + attribute

Resolves  #44713