@GrabYourPitchforks 2020-11-23
I don't understand the context for this PR. What is it attempting to optimize, and how does it accomplish the optimization?