@benaadams 2020-11-23
Calling `GetCustomAttribute` singlar calls `GetCustomAttributes` plural to create an array of attributes which it then throws away returning one item (or throwing if there are more than one item); the optimization is to skip the intermediate array creation.