@benaadams 2020-11-23
> Is this fixing #44713 ?

Ish; though that's `PropertyInfo`, which I haven't done; as its only `RuntimeType`, `RuntimeMethodInfo`

Will add the rest of `MemberInfo`'s derived types `RuntimeEventInfo`, `RuntimeFieldInfo`, `RuntimePropertyInfo`, `RuntimeConstructorInfo` which will then fix that bug.