@benaadams 2020-11-26
Need to investigate why `System.Configuration.ConfigurationManager.Tests` and `System.Text.Json.Tests` are timing out across many variants

Meanwhile some results:

`GetCustomAttribute` which I was originally looking at; brings allocations down to 25% and down to 3% for  #44713; with a performance improvement of 25% and 56% for #44713

|          Method |               Toolchain | Inherit | CustomAttributeVariant |       Mean | Allocated | Ratio |
|---------------- |------------------------ |-------- |----------------------- |-----------:|----------:|------:|
|    PropertyInfo | 	\master\corerun.exe |   False |     GetCustomAttribute |   562.5 ns |     104 B |  1.00 |
|    PropertyInfo |         \pr\corerun.exe |   False |     GetCustomAttribute |   423.8 ns |      24 B |  0.75 |
|                 | 	                    |         |                        |            |           |       |
|            Type | 	\master\corerun.exe |   False |     GetCustomAttribute |   580.1 ns |     104 B |  1.00 |
|            Type |         \pr\corerun.exe |   False |     GetCustomAttribute |   426.5 ns |      24 B |  0.74 |
|                 | 	                    |         |                        |            |           |       |
| ConstructorInfo | 	\master\corerun.exe |   False |     GetCustomAttribute |   569.9 ns |     104 B |  1.00 |
| ConstructorInfo |         \pr\corerun.exe |   False |     GetCustomAttribute |   420.1 ns |      24 B |  0.74 |
|                 | 	                    |         |                        |            |           |       |
|      MethodInfo | 	\master\corerun.exe |   False |     GetCustomAttribute |   610.0 ns |     104 B |  1.00 |
|      MethodInfo |         \pr\corerun.exe |   False |     GetCustomAttribute |   467.4 ns |      24 B |  0.77 |
|                 | 	                    |         |                        |            |           |       |
|       FieldInfo | 	\master\corerun.exe |   False |     GetCustomAttribute |   590.1 ns |     104 B |  1.00 |
|       FieldInfo |         \pr\corerun.exe |   False |     GetCustomAttribute |   440.7 ns |      24 B |  0.75 |
|                 | 	                    |         |                        |            |           |       |
|       EventInfo | 	\master\corerun.exe |   False |     GetCustomAttribute |   583.7 ns |     104 B |  1.00 |
|       EventInfo |         \pr\corerun.exe |   False |     GetCustomAttribute |   420.9 ns |      24 B |  0.72 |


|          Method |               Toolchain | Inherit | CustomAttributeVariant |       Mean | Allocated | Ratio |
|---------------- |------------------------ |-------- |----------------------- |-----------:|----------:|------:|
|    PropertyInfo | 	\master\corerun.exe |    True |     GetCustomAttribute | 1,000.4 ns |     664 B |  1.00 |
|    PropertyInfo |         \pr\corerun.exe |    True |     GetCustomAttribute |   431.2 ns |      24 B |  0.43 |
|                 | 	                    |         |                        |            |           |       |
|            Type | 	\master\corerun.exe |    True |     GetCustomAttribute |   714.6 ns |     104 B |  1.00 |
|            Type |         \pr\corerun.exe |    True |     GetCustomAttribute |   515.3 ns |      24 B |  0.72 |
|                 | 	                    |         |                        |            |           |       |
| ConstructorInfo | 	\master\corerun.exe |    True |     GetCustomAttribute |   576.3 ns |     104 B |  1.00 |
| ConstructorInfo |         \pr\corerun.exe |    True |     GetCustomAttribute |   419.7 ns |      24 B |  0.73 |
|                 | 	                    |         |                        |            |           |       |
|      MethodInfo | 	\master\corerun.exe |    True |     GetCustomAttribute |   611.6 ns |     104 B |  1.00 |
|      MethodInfo |         \pr\corerun.exe |    True |     GetCustomAttribute |   462.9 ns |      24 B |  0.76 |
|                 | 	                    |         |                        |            |           |       |
|       FieldInfo | 	\master\corerun.exe |    True |     GetCustomAttribute |   618.4 ns |     104 B |  1.00 |
|       FieldInfo |         \pr\corerun.exe |    True |     GetCustomAttribute |   436.7 ns |      24 B |  0.71 |
|                 | 	                    |         |                        |            |           |       |
|       EventInfo | 	\master\corerun.exe |    True |     GetCustomAttribute | 1,034.5 ns |     664 B |  1.00 |
|       EventInfo |         \pr\corerun.exe |    True |     GetCustomAttribute |   471.0 ns |      24 B |  0.46 |